launch options size, dynamic bin name
launch options size is now dynamic between the x64/x86 VM, bin file name is no longer hardcoded.

+++ launch_options.hpp
@@ -11,6 +11,6 @@ namespace options
 			bool debug_mode: 1;
 		};
 
-		std::uint64_t options;
+		std::uintptr_t options;
 	};
 }
